requirements.txt
functorch
numpy>=1.15
# torch>=1.8
scipy>=1.2
opt-einsum>=3.3.0
